requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.4.0
requests==2.22.0
azure.identity==1.4.1
azure-keyvault-secrets==4.2.0
azure-core==1.8.2
azure-common==1.1.25
pandas==1.1.3
numpy